Add initial desugar impl for transactions
