+++ tests/jballerina-unit-test/src/test/resources/test-src/statements/transaction/test.bal
@@ -1,8 +1,10 @@
 import ballerina/io;
-import ballerina/transactions as trx;
 
 function testRollback() returns string|error {
-    string|error x =  trap desugaredCode(0, true);
+    //string|error x =  trap desugaredCode(0, true);
+
+    string|error x =  trap actualCode(0, true);
+
     io:println(x);
     return x;
 }
@@ -20,32 +22,33 @@ function testPanic() returns string|error {
 }
 
 
-//function actualCode(int failureCutOff, boolean requestRollback) returns (string) {
-//    string a = "";
-//    a = a + "start";
-//    a = a + " fc-" + failureCutOff.toString();
-//    int count = 0;
-//
-//    transaction {
-//        a = a + " inTrx";
-//        count = count + 1;
-//        if (count <= failureCutOff) {
-//            a = a + " blowUp"; // transaction block panic scenario, Set failure cutoff to 0, for not blowing up.
-//            int bV = blowUp();
-//        }
-//        if (requestRollback) { // Set requestRollback to true if you want to try rollback scenario, otherwise commit
-//            a = a + " Rollback";
-//            rollback;
-//        } else {
-//            a = a + " Commit";
-//            commit;
-//        }
-//        a = a + " endTrx";
-//        a = (a + " end");
-//    }
-//
-//    return a;
-//}
+function actualCode(int failureCutOff, boolean requestRollback) returns (string) {
+    string a = "";
+    a = a + "start";
+    a = a + " fc-" + failureCutOff.toString();
+    int count = 0;
+
+    transaction {
+        a = a + " inTrx";
+        var i = commit;
+        //count = count + 1;
+        //if (count <= failureCutOff) {
+        //    a = a + " blowUp"; // transaction block panic scenario, Set failure cutoff to 0, for not blowing up.
+        //    int bV = blowUp();
+        //}
+        //if (requestRollback) { // Set requestRollback to true if you want to try rollback scenario, otherwise commit
+        //    a = a + " Rollback";
+        //    rollback;
+        //} else {
+        //    a = a + " Commit";
+        //    var i = commit;
+        //}
+        //a = a + " endTrx";
+        //a = (a + " end");
+    }
+
+    return a;
+}
 
 function desugaredCode(int failureCutOff, boolean requestRollBack) returns (string) {
     string a = "";
@@ -53,31 +56,24 @@ function desugaredCode(int failureCutOff, boolean requestRollBack) returns (stri
     a = a + " fc-" + failureCutOff.toString();
     int count = 0;
 
-    string transactionId = "";
     string transactionBlockId = "";
-    trx:TransactionContext|error txnContext =  trx:beginTransaction((), transactionBlockId, "", trx:TWO_PHASE_COMMIT);
-    if (txnContext is error) {
-        panic txnContext;
-    } else {
-        transactionId = txnContext.transactionId;
-        trx:setTransactionContext(txnContext);
-    }
+    string transactionId = transactions:startTransaction(transactionBlockId);
 
     var trxFunc = function () returns error? {
-        var zz = function(trx:Info? info, error? cause, boolean willTry) {
+        var zz = function(transactions:Info? info, error? cause, boolean willTry) {
             io:println("######----RollbackedZZ-----#####");
         };
-        trx:onRollback(zz);
+        transactions:onRollback(zz);
 
-        var xx = function (trx:Info? info) {
+        var xx = function (transactions:Info? info) {
             io:println("#########----CommitedXX----###########");
         };
-        var yy = function (trx:Info? info) {
+        var yy = function (transactions:Info? info) {
             io:println("#########----CommitedYY----###########");
         };
-        trx:onCommit(xx);
-        trx:onCommit(xx);
-        trx:onCommit(yy);
+        transactions:onCommit(xx);
+        transactions:onCommit(xx);
+        transactions:onCommit(yy);
 
         a = a + " inTrx";
         count = count + 1;
@@ -89,7 +85,7 @@ function desugaredCode(int failureCutOff, boolean requestRollBack) returns (stri
 
         if (requestRollBack) {
             a = a + " Rollback";
-            error? rollbackResult = trap trx:rollbackTransaction(transactionBlockId);
+            error? rollbackResult = trap transactions:rollbackTransaction(transactionBlockId);
             if (rollbackResult is error) {
                 io:println("rollback failed ",rollbackResult.reason());
             } else {
@@ -97,11 +93,11 @@ function desugaredCode(int failureCutOff, boolean requestRollBack) returns (stri
             }
         } else {
             a = a + " Commit";
-            boolean isFailed = trx:getAndClearFailure();
+            boolean isFailed = transactions:getAndClearFailure();
             if (!isFailed) {
-                var endSuccess = trap trx:endTransaction(transactionId, transactionBlockId);
+                var endSuccess = trap transactions:endTransaction(transactionId, transactionBlockId);
                 if (endSuccess is string) {
-                    if (endSuccess == trx:OUTCOME_COMMITTED) {
+                    if (endSuccess == transactions:OUTCOME_COMMITTED) {
                         io:println("Commit success");
                     }
                 }
